feat(sync_manager): Turn the single file into a sync_manager/ module with various sub-modules
