+++ crates/nx-core/src/sync_manager/manager.rs
@@ -1,304 +1,31 @@
-use std::collections::{HashMap, HashSet, VecDeque};
-use std::sync::{
-    Arc,
-    atomic::{AtomicU64, Ordering},
-};
-use std::time::{Duration, Instant as StdInstant};
+use std::collections::HashMap;
+use std::sync::{Arc, atomic::AtomicU64};
 
-use nx_net::{NetError, Node, NodeConfig, NodeEvent};
+use nx_net::{Node, NodeConfig};
 use nx_store::Store as NxStore;
-use nx_sync::{GCounter, LwwMap, LwwRegister, NodeId, ORSet, Op, OpKind, PNCounter, Rga};
+use nx_sync::{GCounter, LwwMap, LwwRegister, NodeId, ORSet, Op, PNCounter, Rga};
 use tokio::sync::{RwLock, mpsc, watch};
 use tokio::task::JoinHandle;
-use tracing::{debug, error, info, warn};
+use tracing::{debug, info, warn};
 
 use crate::observability::RuntimeMetrics;
 use crate::sync_config::SyncConfig;
 
-/// Upper bound on how many ops we coalesce into a single PushOps message.
-const BROADCAST_BATCH_MAX: usize = 1024;
-const BROADCAST_COALESCE_DELAY: Duration = Duration::from_millis(1);
-const GCOUNTER_STORE_PREFIX: &str = "__nx/crdt/gcounter/";
-const GCOUNTER_STATE_STORE_PREFIX: &str = "__nx/crdt/state/gcounter/";
-const PNCOUNTER_STORE_PREFIX: &str = "__nx/crdt/pncounter/";
-const PNCOUNTER_STATE_STORE_PREFIX: &str = "__nx/crdt/state/pncounter/";
-const LWW_REGISTER_STORE_PREFIX: &str = "__nx/crdt/lww-register/";
-const LWW_REGISTER_STATE_STORE_PREFIX: &str = "__nx/crdt/state/lww-register/";
-const LWW_MAP_STORE_PREFIX: &str = "__nx/crdt/lww-map/";
-const LWW_MAP_STATE_STORE_PREFIX: &str = "__nx/crdt/state/lww-map/";
-const ORSET_STORE_PREFIX: &str = "__nx/crdt/orset/";
-const ORSET_STATE_STORE_PREFIX: &str = "__nx/crdt/state/orset/";
-const RGA_STORE_PREFIX: &str = "__nx/crdt/rga/";
-const RGA_STATE_STORE_PREFIX: &str = "__nx/crdt/state/rga/";
-const SEEN_OP_STORE_PREFIX: &str = "__nx/crdt/seen-op/";
-const OP_LOG_STORE_PREFIX: &str = "__nx/crdt/op-log/";
-
-#[derive(Debug, Clone, Copy)]
-enum CrdtStoreNamespace {
-    Materialized,
-    State,
-}
-
-#[derive(Debug, Clone, Copy)]
-enum CrdtKind {
-    GCounter,
-    PNCounter,
-    LwwRegister,
-    LwwMap,
-    ORSet,
-    Rga,
-}
-
-impl CrdtKind {
-    fn label(self) -> &'static str {
-        match self {
-            Self::GCounter => "GCounter",
-            Self::PNCounter => "PNCounter",
-            Self::LwwRegister => "LWW-Register",
-            Self::LwwMap => "LWW-Map",
-            Self::ORSet => "ORSet",
-            Self::Rga => "RGA",
-        }
-    }
-
-    fn prefix(self, namespace: CrdtStoreNamespace) -> &'static str {
-        match (self, namespace) {
-            (Self::GCounter, CrdtStoreNamespace::Materialized) => GCOUNTER_STORE_PREFIX,
-            (Self::GCounter, CrdtStoreNamespace::State) => GCOUNTER_STATE_STORE_PREFIX,
-            (Self::PNCounter, CrdtStoreNamespace::Materialized) => PNCOUNTER_STORE_PREFIX,
-            (Self::PNCounter, CrdtStoreNamespace::State) => PNCOUNTER_STATE_STORE_PREFIX,
-            (Self::LwwRegister, CrdtStoreNamespace::Materialized) => LWW_REGISTER_STORE_PREFIX,
-            (Self::LwwRegister, CrdtStoreNamespace::State) => LWW_REGISTER_STATE_STORE_PREFIX,
-            (Self::LwwMap, CrdtStoreNamespace::Materialized) => LWW_MAP_STORE_PREFIX,
-            (Self::LwwMap, CrdtStoreNamespace::State) => LWW_MAP_STATE_STORE_PREFIX,
-            (Self::ORSet, CrdtStoreNamespace::Materialized) => ORSET_STORE_PREFIX,
-            (Self::ORSet, CrdtStoreNamespace::State) => ORSET_STATE_STORE_PREFIX,
-            (Self::Rga, CrdtStoreNamespace::Materialized) => RGA_STORE_PREFIX,
-            (Self::Rga, CrdtStoreNamespace::State) => RGA_STATE_STORE_PREFIX,
-        }
-    }
-}
-
-#[derive(Debug, PartialEq, Eq)]
-enum SeenOpsInsert {
-    AlreadySeen,
-    Inserted { evicted: Vec<String> },
-}
-
-/// Health state for a configured peer.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum PeerHealthState {
-    Healthy,
-    Suspect,
-    Dead,
-}
-
-#[derive(Debug, Clone, Copy)]
-struct PeerHealth {
-    state: PeerHealthState,
-    consecutive_failures: u32,
-}
-
-impl Default for PeerHealth {
-    fn default() -> Self {
-        Self {
-            state: PeerHealthState::Suspect,
-            consecutive_failures: 0,
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-struct SeenOps {
-    ids: HashSet<String>,
-    order: VecDeque<String>,
-    limit: usize,
-}
-
-impl SeenOps {
-    fn new(limit: usize) -> Self {
-        Self {
-            ids: HashSet::new(),
-            order: VecDeque::new(),
-            limit: normalize_seen_ops_limit(limit),
-        }
-    }
-
-    #[cfg(test)]
-    fn contains(&self, op_id: &str) -> bool {
-        self.ids.contains(op_id)
-    }
-
-    fn insert(&mut self, op_id: impl Into<String>) -> SeenOpsInsert {
-        let op_id = op_id.into();
-        if !self.ids.insert(op_id.clone()) {
-            return SeenOpsInsert::AlreadySeen;
-        }
-
-        self.order.push_back(op_id);
-        let mut evicted_ids = Vec::new();
-        while self.order.len() > self.limit {
-            if let Some(evicted) = self.order.pop_front() {
-                self.ids.remove(&evicted);
-                evicted_ids.push(evicted);
-            }
-        }
-        SeenOpsInsert::Inserted {
-            evicted: evicted_ids,
-        }
-    }
-
-    fn len(&self) -> usize {
-        self.ids.len()
-    }
-}
-
-struct ReconnectLoopContext {
-    node: Arc<Node>,
-    peers: Vec<String>,
-    max_peers: usize,
-    initial_delay: Duration,
-    max_delay: Duration,
-    peer_dead_after_failures: u32,
-    shutdown_rx: watch::Receiver<bool>,
-    metrics: Arc<RuntimeMetrics>,
-    peer_health: Arc<RwLock<HashMap<String, PeerHealth>>>,
-}
-
-struct AntiEntropyLoopContext {
-    node: Arc<Node>,
-    peers: Vec<String>,
-    interval: Duration,
-    shutdown_rx: watch::Receiver<bool>,
-    metrics: Arc<RuntimeMetrics>,
-}
-
-struct BroadcastLoopContext {
-    node: Arc<Node>,
-    seen_ops: Arc<RwLock<SeenOps>>,
-    seen_ops_next_sequence: Arc<AtomicU64>,
-    op_log: Arc<RwLock<Vec<Op>>>,
-    op_log_next_sequence: Arc<AtomicU64>,
-    op_log_limit: usize,
-    store: Arc<NxStore>,
-    metrics: Arc<RuntimeMetrics>,
-}
-
-struct RemoteOpApplyContext<'a> {
-    counters: &'a Arc<RwLock<HashMap<String, GCounter>>>,
-    pncounters: &'a Arc<RwLock<HashMap<String, PNCounter>>>,
-    lww_registers: &'a Arc<RwLock<HashMap<String, LwwRegister>>>,
-    lww_maps: &'a Arc<RwLock<HashMap<String, LwwMap>>>,
-    orsets: &'a Arc<RwLock<HashMap<String, ORSet>>>,
-    rgas: &'a Arc<RwLock<HashMap<String, Rga>>>,
-    seen_ops: &'a Arc<RwLock<SeenOps>>,
-    seen_ops_next_sequence: &'a Arc<AtomicU64>,
-    op_log: &'a Arc<RwLock<Vec<Op>>>,
-    op_log_next_sequence: &'a Arc<AtomicU64>,
-    op_log_limit: usize,
-    store: &'a Arc<NxStore>,
-    metrics: &'a Arc<RuntimeMetrics>,
-}
-
-struct RemoteCrdtRegistries<'a> {
-    counters: &'a HashMap<String, GCounter>,
-    pncounters: &'a HashMap<String, PNCounter>,
-    lww_registers: &'a HashMap<String, LwwRegister>,
-    lww_maps: &'a HashMap<String, LwwMap>,
-    orsets: &'a HashMap<String, ORSet>,
-    rgas: &'a HashMap<String, Rga>,
-}
-
-struct RemoteCrdtUpdates<'a> {
-    counters: &'a mut HashMap<String, GCounter>,
-    pncounters: &'a mut HashMap<String, PNCounter>,
-    lww_registers: &'a mut HashMap<String, LwwRegister>,
-    lww_maps: &'a mut HashMap<String, LwwMap>,
-    orsets: &'a mut HashMap<String, ORSet>,
-    rgas: &'a mut HashMap<String, Rga>,
-}
-
-struct RemoteCrdtUpdateBatch<'a> {
-    counters: &'a HashMap<String, GCounter>,
-    pncounters: &'a HashMap<String, PNCounter>,
-    lww_registers: &'a HashMap<String, LwwRegister>,
-    lww_maps: &'a HashMap<String, LwwMap>,
-    orsets: &'a HashMap<String, ORSet>,
-    rgas: &'a HashMap<String, Rga>,
-}
-
-struct OpPersistencePlan {
-    op: Op,
-    seen_sequence: u64,
-    op_log_sequence: u64,
-    seen_evicted: Vec<String>,
-    op_log_evicted: Vec<Op>,
-}
-
-#[derive(Clone)]
-struct NodeEventContext {
-    counters: Arc<RwLock<HashMap<String, GCounter>>>,
-    pncounters: Arc<RwLock<HashMap<String, PNCounter>>>,
-    lww_registers: Arc<RwLock<HashMap<String, LwwRegister>>>,
-    lww_maps: Arc<RwLock<HashMap<String, LwwMap>>>,
-    orsets: Arc<RwLock<HashMap<String, ORSet>>>,
-    rgas: Arc<RwLock<HashMap<String, Rga>>>,
-    seen_ops: Arc<RwLock<SeenOps>>,
-    seen_ops_next_sequence: Arc<AtomicU64>,
-    op_log: Arc<RwLock<Vec<Op>>>,
-    op_log_next_sequence: Arc<AtomicU64>,
-    op_log_limit: usize,
-    store: Arc<NxStore>,
-    metrics: Arc<RuntimeMetrics>,
-    node: Arc<Node>,
-    peer_health: Arc<RwLock<HashMap<String, PeerHealth>>>,
-    peer_node_ids: Arc<RwLock<HashMap<String, NodeId>>>,
-    anti_entropy_watermarks: Arc<RwLock<HashMap<NodeId, String>>>,
-    peer_dead_after_failures: u32,
-}
-
-struct PeerReconnectState {
-    addr: String,
-    delay: Duration,
-    next_attempt_at: StdInstant,
-}
-
-impl PeerReconnectState {
-    fn new(addr: String, initial_delay: Duration, now: StdInstant) -> Self {
-        Self {
-            addr,
-            delay: initial_delay,
-            next_attempt_at: now,
-        }
-    }
-
-    fn reset(&mut self, initial_delay: Duration, now: StdInstant) {
-        self.delay = initial_delay;
-        self.next_attempt_at = now;
-    }
-
-    fn record_failure(&mut self, max_delay: Duration, now: StdInstant) -> Duration {
-        let attempt_delay = self.delay;
-        self.next_attempt_at = now + attempt_delay;
-        self.delay = next_reconnect_delay(attempt_delay, max_delay);
-        attempt_delay
-    }
-}
-
-struct ConfiguredPeerConnectContext<'a> {
-    node: &'a Node,
-    max_peers: usize,
-    peer_dead_after_failures: u32,
-    metrics: &'a Arc<RuntimeMetrics>,
-    peer_health: &'a Arc<RwLock<HashMap<String, PeerHealth>>>,
-}
-
-enum ConfiguredPeerConnectOutcome {
-    Connected,
-    AlreadyConnected,
-    SlotLimitReached,
-    Failed,
-}
+use super::peer::{
+    ConfiguredPeerConnectContext, ConfiguredPeerConnectOutcome, PeerHealth, PeerHealthState,
+    normalize_peer_dead_after_failures,
+};
+use super::replication::{
+    drain_node_events, handle_node_event, normalize_op_log_limit, normalize_seen_ops_limit,
+    spawn_anti_entropy_loop, spawn_broadcast_loop, spawn_reconnect_loop,
+    try_connect_configured_peer,
+};
+use super::storage::{
+    hydrate_gcounter_registry, hydrate_lww_map_registry, hydrate_lww_register_registry,
+    hydrate_op_log, hydrate_orset_registry, hydrate_pncounter_registry, hydrate_rga_registry,
+    hydrate_seen_ops,
+};
+use super::types::*;
 
 /// Clonable, cheap handle to the SyncManager exposed to host calls.
 #[derive(Clone)]
@@ -842,1945 +569,16 @@ impl SyncManager {
     }
 }
 
-/// Spawn the outbound broadcast drain loop.
-fn spawn_broadcast_loop(
-    context: BroadcastLoopContext,
-    mut op_rx: mpsc::Receiver<Op>,
-    mut shutdown_rx: watch::Receiver<bool>,
-) -> JoinHandle<()> {
-    tokio::spawn(async move {
-        loop {
-            tokio::select! {
-                _ = shutdown_rx.changed() => {
-                    if *shutdown_rx.borrow() {
-                        debug!("broadcast loop shutdown requested");
-                        drain_broadcast_queue(
-                            &context,
-                            &mut op_rx,
-                        ).await;
-                        break;
-                    }
-                }
-                op = op_rx.recv() => {
-                    let Some(first) = op else {
-                        break;
-                    };
-                    broadcast_batch(
-                        &context,
-                        &mut op_rx,
-                        first,
-                    ).await;
-                }
-            }
-        }
-        debug!("broadcast loop terminated");
-    })
-}
-
-async fn try_connect_configured_peer(
-    context: &ConfiguredPeerConnectContext<'_>,
-    peer_addr: &str,
-) -> ConfiguredPeerConnectOutcome {
-    if context.node.is_connected_addr(peer_addr).await {
-        mark_peer_success(context.peer_health, peer_addr).await;
-        return ConfiguredPeerConnectOutcome::AlreadyConnected;
-    }
-
-    if context.node.connected_peer_count().await >= context.max_peers {
-        debug!(
-            peer = %peer_addr,
-            limit = context.max_peers,
-            "skipping configured peer connect: peer slot limit reached"
-        );
-        return ConfiguredPeerConnectOutcome::SlotLimitReached;
-    }
-
-    match context.node.connect_to_peer(peer_addr).await {
-        Ok(()) => {
-            mark_peer_success(context.peer_health, peer_addr).await;
-            ConfiguredPeerConnectOutcome::Connected
-        }
-        Err(NetError::PeerLimitReached(limit)) => {
-            debug!(
-                peer = %peer_addr,
-                limit,
-                "skipping configured peer connect: peer slot limit reached"
-            );
-            ConfiguredPeerConnectOutcome::SlotLimitReached
-        }
-        Err(e) => {
-            context.metrics.record_sync_error();
-            mark_peer_failure(
-                context.peer_health,
-                peer_addr,
-                context.peer_dead_after_failures,
-            )
-            .await;
-            debug!(peer = %peer_addr, error = %e, "configured peer connect failed");
-            ConfiguredPeerConnectOutcome::Failed
-        }
-    }
-}
-
-fn spawn_reconnect_loop(context: ReconnectLoopContext) -> Option<JoinHandle<()>> {
-    let ReconnectLoopContext {
-        node,
-        peers,
-        max_peers,
-        initial_delay,
-        max_delay,
-        peer_dead_after_failures,
-        mut shutdown_rx,
-        metrics,
-        peer_health,
-    } = context;
-
-    if peers.is_empty() {
-        return None;
-    }
-
-    Some(tokio::spawn(async move {
-        let initial_delay = normalize_reconnect_delay(initial_delay);
-        let max_delay = max_delay.max(initial_delay);
-        let peer_dead_after_failures = normalize_peer_dead_after_failures(peer_dead_after_failures);
-        let now = StdInstant::now();
-        let mut state = peers
-            .into_iter()
-            .map(|addr| PeerReconnectState::new(addr, initial_delay, now))
-            .collect::<Vec<_>>();
-
-        loop {
-            let mut sleep_for: Option<Duration> = None;
-            let now = StdInstant::now();
-            let connect_context = ConfiguredPeerConnectContext {
-                node: node.as_ref(),
-                max_peers,
-                peer_dead_after_failures,
-                metrics: &metrics,
-                peer_health: &peer_health,
-            };
-
-            for peer in state.iter_mut() {
-                let peer_addr = peer.addr.as_str();
-
-                if node.is_connected_addr(peer_addr).await {
-                    mark_peer_success(&peer_health, peer_addr).await;
-                    peer.reset(initial_delay, StdInstant::now());
-                    continue;
-                }
-
-                if let Some(wait) = peer.next_attempt_at.checked_duration_since(now)
-                    && !wait.is_zero()
-                {
-                    sleep_for = Some(sleep_for.map_or(wait, |current| current.min(wait)));
-                    continue;
-                }
-
-                match try_connect_configured_peer(&connect_context, peer_addr).await {
-                    ConfiguredPeerConnectOutcome::Connected => {
-                        info!(peer = %peer_addr, "reconnected configured peer");
-                        peer.reset(initial_delay, StdInstant::now());
-                    }
-                    ConfiguredPeerConnectOutcome::AlreadyConnected => {
-                        peer.reset(initial_delay, StdInstant::now());
-                    }
-                    ConfiguredPeerConnectOutcome::SlotLimitReached => {
-                        break;
-                    }
-                    ConfiguredPeerConnectOutcome::Failed => {
-                        let attempt_delay = peer.record_failure(max_delay, StdInstant::now());
-                        sleep_for = Some(
-                            sleep_for.map_or(attempt_delay, |current| current.min(attempt_delay)),
-                        );
-                    }
-                }
-            }
-            let sleep_for = sleep_for.unwrap_or(initial_delay);
-
-            tokio::select! {
-                _ = shutdown_rx.changed() => {
-                    if *shutdown_rx.borrow() {
-                        debug!("reconnect loop shutdown requested");
-                        break;
-                    }
-                }
-                _ = tokio::time::sleep(sleep_for) => {}
-            }
-        }
-        debug!("reconnect loop terminated");
-    }))
-}
-
-fn spawn_anti_entropy_loop(context: AntiEntropyLoopContext) -> Option<JoinHandle<()>> {
-    let AntiEntropyLoopContext {
-        node,
-        peers,
-        interval,
-        mut shutdown_rx,
-        metrics,
-    } = context;
-
-    if peers.is_empty() {
-        return None;
-    }
-
-    Some(tokio::spawn(async move {
-        let interval = normalize_anti_entropy_interval(interval);
-        loop {
-            tokio::select! {
-                _ = shutdown_rx.changed() => {
-                    if *shutdown_rx.borrow() {
-                        debug!("anti-entropy loop shutdown requested");
-                        break;
-                    }
-                }
-                _ = tokio::time::sleep(interval) => {
-                    for peer in &peers {
-                        if !node.is_connected_addr(peer).await {
-                            continue;
-                        }
-
-                        // A single "last seen OpId" is not a safe causal frontier: a peer can
-                        // receive a newer op while an older broadcast was dropped. Until the
-                        // protocol has contiguous/causal metadata, anti-entropy pulls the bounded
-                        // op-log and relies on OpId deduplication on the receiver.
-                        if let Err(e) = node.send_pull_since_to_addr(peer, None).await {
-                            metrics.record_sync_error();
-                            debug!(peer = %peer, error = %e, "anti-entropy pull failed");
-                        } else {
-                            debug!(peer = %peer, "anti-entropy pull requested");
-                        }
-                    }
-                }
-            }
-        }
-        debug!("anti-entropy loop terminated");
-    }))
-}
-
-fn normalize_reconnect_delay(delay: Duration) -> Duration {
-    delay.max(Duration::from_millis(1))
-}
-
-fn normalize_anti_entropy_interval(interval: Duration) -> Duration {
-    interval.max(Duration::from_millis(1))
-}
-
-fn normalize_op_log_limit(limit: usize) -> usize {
-    limit.max(1)
-}
-
-fn normalize_seen_ops_limit(limit: usize) -> usize {
-    limit.max(1)
-}
-
-fn next_reconnect_delay(current: Duration, max_delay: Duration) -> Duration {
-    current.saturating_mul(2).min(max_delay.max(current))
-}
-
-fn normalize_peer_dead_after_failures(failures: u32) -> u32 {
-    failures.max(1)
-}
-
-fn record_peer_success(peer_health: &mut HashMap<String, PeerHealth>, peer: &str) {
-    let health = peer_health.entry(peer.to_string()).or_default();
-    health.state = PeerHealthState::Healthy;
-    health.consecutive_failures = 0;
-}
-
-fn record_peer_failure(
-    peer_health: &mut HashMap<String, PeerHealth>,
-    peer: &str,
-    dead_after_failures: u32,
-) {
-    let dead_after_failures = normalize_peer_dead_after_failures(dead_after_failures);
-    let health = peer_health.entry(peer.to_string()).or_default();
-    health.consecutive_failures = health.consecutive_failures.saturating_add(1);
-    health.state = if health.consecutive_failures >= dead_after_failures {
-        PeerHealthState::Dead
-    } else {
-        PeerHealthState::Suspect
-    };
-}
-
-async fn mark_peer_success(peer_health: &Arc<RwLock<HashMap<String, PeerHealth>>>, peer: &str) {
-    let mut peer_health = peer_health.write().await;
-    record_peer_success(&mut peer_health, peer);
-}
-
-async fn mark_peer_failure(
-    peer_health: &Arc<RwLock<HashMap<String, PeerHealth>>>,
-    peer: &str,
-    dead_after_failures: u32,
-) {
-    let mut peer_health = peer_health.write().await;
-    record_peer_failure(&mut peer_health, peer, dead_after_failures);
-}
-
-async fn mark_known_peer_success(
-    peer_health: &Arc<RwLock<HashMap<String, PeerHealth>>>,
-    peer: &str,
-) {
-    let mut peer_health = peer_health.write().await;
-    if peer_health.contains_key(peer) {
-        record_peer_success(&mut peer_health, peer);
-    }
-}
-
-async fn mark_known_peer_failure(
-    peer_health: &Arc<RwLock<HashMap<String, PeerHealth>>>,
-    peer: &str,
-    dead_after_failures: u32,
-) {
-    let mut peer_health = peer_health.write().await;
-    if peer_health.contains_key(peer) {
-        record_peer_failure(&mut peer_health, peer, dead_after_failures);
-    }
-}
-
-async fn broadcast_batch(
-    context: &BroadcastLoopContext,
-    op_rx: &mut mpsc::Receiver<Op>,
-    first: Op,
-) {
-    let mut batch = Vec::with_capacity(BROADCAST_BATCH_MAX);
-    batch.push(first);
-
-    // Give producers a tiny window to fill the channel so sustained load is
-    // sent in larger PushOps messages instead of many tiny network writes.
-    tokio::time::sleep(BROADCAST_COALESCE_DELAY).await;
-
-    while batch.len() < BROADCAST_BATCH_MAX {
-        match op_rx.try_recv() {
-            Ok(op) => batch.push(op),
-            Err(_) => break,
-        }
-    }
-
-    let count = batch.len();
-    if let Err(e) = remember_ops(
-        &context.seen_ops,
-        &context.seen_ops_next_sequence,
-        &context.op_log,
-        &context.op_log_next_sequence,
-        context.op_log_limit,
-        &context.store,
-        &batch,
-    )
-    .await
-    {
-        context.metrics.record_sync_error();
-        warn!(error = %e, "failed to persist local op dedup metadata");
-        return;
-    }
-    let started = std::time::Instant::now();
-    if let Err(e) = context.node.broadcast_ops(batch).await {
-        context.metrics.record_sync_error();
-        warn!(error = %e, count, "broadcast partially failed; ops dropped for failed peers");
-    } else {
-        context.metrics.record_broadcast_batch(count);
-        context.metrics.record_sync_latency(started.elapsed());
-        debug!(count, "broadcast batch sent");
-    }
-}
-
-async fn drain_broadcast_queue(context: &BroadcastLoopContext, op_rx: &mut mpsc::Receiver<Op>) {
-    while let Ok(first) = op_rx.try_recv() {
-        broadcast_batch(context, op_rx, first).await;
-    }
-}
-
-async fn remember_ops(
-    seen_ops: &Arc<RwLock<SeenOps>>,
-    seen_ops_next_sequence: &Arc<AtomicU64>,
-    op_log: &Arc<RwLock<Vec<Op>>>,
-    op_log_next_sequence: &Arc<AtomicU64>,
-    op_log_limit: usize,
-    store: &Arc<NxStore>,
-    ops: &[Op],
-) -> anyhow::Result<()> {
-    let mut seen = seen_ops.write().await;
-    let mut log = op_log.write().await;
-    let mut next_seen_sequence = seen_ops_next_sequence.load(Ordering::Relaxed);
-    let mut next_op_log_sequence = op_log_next_sequence.load(Ordering::Relaxed);
-    let mut inserted_ops = Vec::new();
-    let mut inserted_ids = Vec::new();
-    let mut inserts = Vec::new();
-    for op in ops {
-        let op_id = op.id.as_str();
-        if !seen.ids.contains(op_id) && !inserted_ids.iter().any(|id| id == op_id) {
-            inserted_ids.push(op_id.to_string());
-            inserted_ops.push(op.clone());
-        }
-    }
-
-    let seen_evicted = plan_seen_evictions(&seen, &inserted_ids);
-    let op_log_evicted = plan_op_log_evictions(&log, inserted_ops.len(), op_log_limit);
-    let last_insert_index = inserted_ops.len().saturating_sub(1);
-
-    for (index, op) in inserted_ops.iter().enumerate() {
-        inserts.push(OpPersistencePlan {
-            op: op.clone(),
-            seen_sequence: next_seen_sequence,
-            op_log_sequence: next_op_log_sequence,
-            seen_evicted: if index == last_insert_index {
-                seen_evicted.clone()
-            } else {
-                Vec::new()
-            },
-            op_log_evicted: if index == last_insert_index {
-                op_log_evicted.clone()
-            } else {
-                Vec::new()
-            },
-        });
-        next_seen_sequence = next_seen_sequence.saturating_add(1);
-        next_op_log_sequence = next_op_log_sequence.saturating_add(1);
-    }
-
-    persist_local_ops_batch(store, &inserts)?;
-    apply_seen_insertions(&mut seen, &inserted_ids, &seen_evicted);
-    log.extend(inserted_ops);
-    prune_op_log_and_return_evicted(&mut log, op_log_limit);
-    seen_ops_next_sequence.store(next_seen_sequence, Ordering::Relaxed);
-    op_log_next_sequence.store(next_op_log_sequence, Ordering::Relaxed);
-    Ok(())
-}
-
-fn plan_seen_evictions(seen: &SeenOps, inserted_ids: &[String]) -> Vec<String> {
-    let total_len = seen.order.len().saturating_add(inserted_ids.len());
-    let remove_count = total_len.saturating_sub(seen.limit);
-    if remove_count == 0 {
-        return Vec::new();
-    }
-
-    let mut evicted = seen
-        .order
-        .iter()
-        .take(remove_count)
-        .cloned()
-        .collect::<Vec<_>>();
-    if remove_count > seen.order.len() {
-        evicted.extend(
-            inserted_ids
-                .iter()
-                .take(remove_count - seen.order.len())
-                .cloned(),
-        );
-    }
-    evicted
-}
-
-fn apply_seen_insertions(seen: &mut SeenOps, inserted_ids: &[String], evicted: &[String]) {
-    for op_id in inserted_ids {
-        seen.ids.insert(op_id.clone());
-        seen.order.push_back(op_id.clone());
-    }
-    for op_id in evicted {
-        seen.ids.remove(op_id);
-        let _ = seen.order.pop_front();
-    }
-}
-
-fn plan_op_log_evictions(log: &[Op], inserted_count: usize, limit: usize) -> Vec<Op> {
-    let limit = normalize_op_log_limit(limit);
-    let total_len = log.len().saturating_add(inserted_count);
-    let remove_count = total_len.saturating_sub(limit);
-    log.iter().take(remove_count).cloned().collect()
-}
-
-fn prune_op_log_and_return_evicted(log: &mut Vec<Op>, limit: usize) -> Vec<Op> {
-    let limit = normalize_op_log_limit(limit);
-    if log.len() > limit {
-        let remove_count = log.len() - limit;
-        log.drain(..remove_count).collect()
-    } else {
-        Vec::new()
-    }
-}
-
-async fn op_log_since(op_log: &Arc<RwLock<Vec<Op>>>, since_op_id: Option<&str>) -> Vec<Op> {
-    let log = op_log.read().await;
-    match since_op_id {
-        Some(op_id) => log
-            .iter()
-            .position(|op| op.id.as_str() == op_id)
-            .map_or_else(|| log.clone(), |index| log[index + 1..].to_vec()),
-        None => log.clone(),
-    }
-}
-
-async fn drain_node_events(event_rx: &mut mpsc::Receiver<NodeEvent>, context: &NodeEventContext) {
-    while let Ok(event) = event_rx.try_recv() {
-        handle_node_event(event, context).await;
-    }
-}
-
-async fn handle_node_event(event: NodeEvent, context: &NodeEventContext) {
-    match event {
-        NodeEvent::OpsReceived { from, ops } => {
-            debug!(from = %from, count = ops.len(), "received ops from peer");
-            let last_received_op_id = ops.last().map(|op| op.id.as_str().to_string());
-            let apply_context = RemoteOpApplyContext {
-                counters: &context.counters,
-                pncounters: &context.pncounters,
-                lww_registers: &context.lww_registers,
-                lww_maps: &context.lww_maps,
-                orsets: &context.orsets,
-                rgas: &context.rgas,
-                seen_ops: &context.seen_ops,
-                seen_ops_next_sequence: &context.seen_ops_next_sequence,
-                op_log: &context.op_log,
-                op_log_next_sequence: &context.op_log_next_sequence,
-                op_log_limit: context.op_log_limit,
-                store: &context.store,
-                metrics: &context.metrics,
-            };
-            if let Err(e) = apply_remote_ops(&ops, &apply_context).await {
-                context.metrics.record_sync_error();
-                error!(error = %e, "failed to apply remote ops batch");
-                return;
-            }
-            if let Some(op_id) = last_received_op_id {
-                context
-                    .anti_entropy_watermarks
-                    .write()
-                    .await
-                    .insert(from, op_id);
-            }
-        }
-        NodeEvent::PullRequested {
-            from,
-            addr,
-            since_op_id,
-        } => {
-            let ops = op_log_since(&context.op_log, since_op_id.as_deref()).await;
-            let count = ops.len();
-            if let Err(e) = context.node.send_ops_to_addr(&addr, ops).await {
-                context.metrics.record_sync_error();
-                warn!(peer = %from, addr = %addr, error = %e, "failed to answer pull request");
-            } else {
-                debug!(peer = %from, addr = %addr, count, "answered pull request");
-            }
-        }
-        NodeEvent::PeerConnected {
-            node_id,
-            addr,
-            peers_connected,
-        } => {
-            mark_known_peer_success(&context.peer_health, &addr).await;
-            context
-                .peer_node_ids
-                .write()
-                .await
-                .insert(addr.clone(), node_id.clone());
-            context.metrics.record_peer_connect();
-            context.metrics.set_peers_connected(peers_connected);
-            info!(peer = %node_id, addr = %addr, "peer connected");
-        }
-        NodeEvent::PeerDisconnected {
-            node_id,
-            addr,
-            peers_connected,
-        } => {
-            mark_known_peer_failure(
-                &context.peer_health,
-                &addr,
-                context.peer_dead_after_failures,
-            )
-            .await;
-            context.peer_node_ids.write().await.remove(&addr);
-            context.metrics.record_peer_disconnect();
-            context.metrics.set_peers_connected(peers_connected);
-            info!(peer = %node_id, addr = %addr, "peer disconnected");
-        }
-    }
-}
-
-pub(crate) fn materialized_gcounter_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::GCounter, CrdtStoreNamespace::Materialized, key)
-}
-
-fn durable_gcounter_state_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::GCounter, CrdtStoreNamespace::State, key)
-}
-
-pub(crate) fn materialized_pncounter_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::PNCounter, CrdtStoreNamespace::Materialized, key)
-}
-
-fn durable_pncounter_state_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::PNCounter, CrdtStoreNamespace::State, key)
-}
-
-pub(crate) fn materialized_lww_register_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::LwwRegister, CrdtStoreNamespace::Materialized, key)
-}
-
-fn durable_lww_register_state_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::LwwRegister, CrdtStoreNamespace::State, key)
-}
-
-pub(crate) fn materialized_lww_map_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::LwwMap, CrdtStoreNamespace::Materialized, key)
-}
-
-fn durable_lww_map_state_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::LwwMap, CrdtStoreNamespace::State, key)
-}
-
-pub(crate) fn materialized_orset_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::ORSet, CrdtStoreNamespace::Materialized, key)
-}
-
-fn durable_orset_state_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::ORSet, CrdtStoreNamespace::State, key)
-}
-
-pub(crate) fn materialized_rga_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::Rga, CrdtStoreNamespace::Materialized, key)
-}
-
-fn durable_rga_state_key(key: &str) -> Vec<u8> {
-    crdt_store_key(CrdtKind::Rga, CrdtStoreNamespace::State, key)
-}
-
-fn seen_op_store_key(op_id: &str) -> Vec<u8> {
-    prefixed_store_key(SEEN_OP_STORE_PREFIX, op_id)
-}
-
-fn op_log_store_key(op_id: &str) -> Vec<u8> {
-    prefixed_store_key(OP_LOG_STORE_PREFIX, op_id)
-}
-
-fn crdt_store_key(kind: CrdtKind, namespace: CrdtStoreNamespace, key: &str) -> Vec<u8> {
-    prefixed_store_key(kind.prefix(namespace), key)
-}
-
-fn prefixed_store_key(prefix: &str, key: &str) -> Vec<u8> {
-    let mut out = Vec::with_capacity(prefix.len() + key.len());
-    out.extend_from_slice(prefix.as_bytes());
-    out.extend_from_slice(key.as_bytes());
-    out
-}
-
-fn logical_gcounter_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(
-        store_key,
-        CrdtKind::GCounter,
-        CrdtStoreNamespace::Materialized,
-    )
-}
-
-fn logical_gcounter_state_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(store_key, CrdtKind::GCounter, CrdtStoreNamespace::State)
-}
-
-fn logical_pncounter_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(
-        store_key,
-        CrdtKind::PNCounter,
-        CrdtStoreNamespace::Materialized,
-    )
-}
-
-fn logical_pncounter_state_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(store_key, CrdtKind::PNCounter, CrdtStoreNamespace::State)
-}
-
-fn logical_lww_register_state_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(store_key, CrdtKind::LwwRegister, CrdtStoreNamespace::State)
-}
-
-fn logical_lww_map_state_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(store_key, CrdtKind::LwwMap, CrdtStoreNamespace::State)
-}
-
-fn logical_orset_state_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(store_key, CrdtKind::ORSet, CrdtStoreNamespace::State)
-}
-
-fn logical_rga_state_key(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_crdt_key(store_key, CrdtKind::Rga, CrdtStoreNamespace::State)
-}
-
-fn logical_seen_op_id(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_key_for_prefix(store_key, SEEN_OP_STORE_PREFIX, "seen OpId")
-}
-
-fn logical_op_log_id(store_key: &[u8]) -> anyhow::Result<String> {
-    logical_key_for_prefix(store_key, OP_LOG_STORE_PREFIX, "op log OpId")
-}
-
-fn logical_crdt_key(
-    store_key: &[u8],
-    kind: CrdtKind,
-    namespace: CrdtStoreNamespace,
-) -> anyhow::Result<String> {
-    let namespace_label = match namespace {
-        CrdtStoreNamespace::Materialized => "materialized",
-        CrdtStoreNamespace::State => "durable state",
-    };
-    logical_key_for_prefix(
-        store_key,
-        kind.prefix(namespace),
-        &format!("{namespace_label} {}", kind.label()),
-    )
-}
-
-fn logical_key_for_prefix(store_key: &[u8], prefix: &str, kind: &str) -> anyhow::Result<String> {
-    let key = store_key
-        .strip_prefix(prefix.as_bytes())
-        .ok_or_else(|| anyhow::anyhow!("invalid {kind} key"))?;
-
-    String::from_utf8(key.to_vec()).map_err(|e| anyhow::anyhow!("invalid UTF-8 in {kind} key: {e}"))
-}
-
-fn parse_seen_op_sequence(bytes: &[u8]) -> anyhow::Result<u64> {
-    let buf: [u8; 8] = bytes.try_into().map_err(|_| {
-        anyhow::anyhow!(
-            "invalid seen OpId sequence length: expected 8, got {}",
-            bytes.len()
-        )
-    })?;
-
-    Ok(u64::from_be_bytes(buf))
-}
-
-fn parse_materialized_gcounter_value(bytes: &[u8]) -> anyhow::Result<u64> {
-    let buf: [u8; 8] = bytes.try_into().map_err(|_| {
-        anyhow::anyhow!(
-            "invalid materialized GCounter value length: expected 8, got {}",
-            bytes.len()
-        )
-    })?;
-
-    Ok(u64::from_le_bytes(buf))
-}
-
-fn parse_materialized_pncounter_value(bytes: &[u8]) -> anyhow::Result<i64> {
-    let buf: [u8; 8] = bytes.try_into().map_err(|_| {
-        anyhow::anyhow!(
-            "invalid materialized PNCounter value length: expected 8, got {}",
-            bytes.len()
-        )
-    })?;
-
-    Ok(i64::from_le_bytes(buf))
-}
-
-fn encode_durable_op_log_value(sequence: u64, op: &Op) -> anyhow::Result<Vec<u8>> {
-    let op_bytes = op.to_bytes()?;
-    let mut out = Vec::with_capacity(8 + op_bytes.len());
-    out.extend_from_slice(&sequence.to_be_bytes());
-    out.extend_from_slice(&op_bytes);
-    Ok(out)
-}
-
-fn parse_durable_op_log_value(bytes: &[u8]) -> anyhow::Result<(u64, Op)> {
-    if bytes.len() < 8 {
-        anyhow::bail!(
-            "invalid op log value length: expected at least 8, got {}",
-            bytes.len()
-        );
-    }
-    let mut seq = [0u8; 8];
-    seq.copy_from_slice(&bytes[..8]);
-    let op = Op::from_bytes(&bytes[8..])
-        .map_err(|e| anyhow::anyhow!("invalid durable op log JSON: {e}"))?;
-    Ok((u64::from_be_bytes(seq), op))
-}
-
-fn hydrate_gcounter_registry(
-    store: &NxStore,
-    node_id: &NodeId,
-    counters: &mut HashMap<String, GCounter>,
-) -> anyhow::Result<usize> {
-    let durable_count = hydrate_durable_gcounter_state(store, counters)?;
-    let materialized_count = hydrate_materialized_gcounter_values(store, node_id, counters)?;
-
-    debug!(
-        durable_count,
-        materialized_count,
-        total = counters.len(),
-        "hydrated GCounter registry from sled"
-    );
-    Ok(counters.len())
-}
-
-fn hydrate_pncounter_registry(
-    store: &NxStore,
-    node_id: &NodeId,
-    counters: &mut HashMap<String, PNCounter>,
-) -> anyhow::Result<usize> {
-    let durable_count = hydrate_durable_pncounter_state(store, counters)?;
-    let materialized_count = hydrate_materialized_pncounter_values(store, node_id, counters)?;
-
-    debug!(
-        durable_count,
-        materialized_count,
-        total = counters.len(),
-        "hydrated PNCounter registry from sled"
-    );
-    Ok(counters.len())
-}
-
-fn hydrate_lww_register_registry(
-    store: &NxStore,
-    registers: &mut HashMap<String, LwwRegister>,
-) -> anyhow::Result<usize> {
-    let durable_count = hydrate_durable_lww_register_state(store, registers)?;
-
-    debug!(
-        durable_count,
-        total = registers.len(),
-        "hydrated LWW-Register registry from sled"
-    );
-    Ok(registers.len())
-}
-
-fn hydrate_lww_map_registry(
-    store: &NxStore,
-    maps: &mut HashMap<String, LwwMap>,
-) -> anyhow::Result<usize> {
-    let durable_count = hydrate_durable_lww_map_state(store, maps)?;
-
-    debug!(
-        durable_count,
-        total = maps.len(),
-        "hydrated LWW-Map registry from sled"
-    );
-    Ok(maps.len())
-}
-
-fn hydrate_orset_registry(
-    store: &NxStore,
-    sets: &mut HashMap<String, ORSet>,
-) -> anyhow::Result<usize> {
-    let durable_count = hydrate_durable_orset_state(store, sets)?;
-
-    debug!(
-        durable_count,
-        total = sets.len(),
-        "hydrated ORSet registry from sled"
-    );
-    Ok(sets.len())
-}
-
-fn hydrate_rga_registry(store: &NxStore, rgas: &mut HashMap<String, Rga>) -> anyhow::Result<usize> {
-    let durable_count = hydrate_durable_rga_state(store, rgas)?;
-
-    debug!(
-        durable_count,
-        total = rgas.len(),
-        "hydrated RGA registry from sled"
-    );
-    Ok(rgas.len())
-}
-
-fn hydrate_seen_ops(store: &NxStore, limit: usize) -> anyhow::Result<(SeenOps, u64)> {
-    let limit = normalize_seen_ops_limit(limit);
-    let entries = store.scan_prefix(SEEN_OP_STORE_PREFIX.as_bytes())?;
-    let mut ordered = Vec::with_capacity(entries.len());
-    let mut next_sequence = 0;
-
-    for (store_key, value_bytes) in entries {
-        let op_id = match logical_seen_op_id(&store_key) {
-            Ok(op_id) => op_id,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid seen OpId key");
-                continue;
-            }
-        };
-        let sequence = match parse_seen_op_sequence(&value_bytes) {
-            Ok(sequence) => sequence,
-            Err(e) => {
-                warn!(op_id = %op_id, error = %e, "skipping invalid seen OpId metadata");
-                continue;
-            }
-        };
-
-        next_sequence = next_sequence.max(sequence.saturating_add(1));
-        ordered.push((sequence, op_id));
-    }
-
-    ordered.sort_by_key(|(sequence, _)| *sequence);
-    let mut seen = SeenOps::new(limit);
-    let mut evicted = Vec::new();
-    for (_, op_id) in ordered {
-        if let SeenOpsInsert::Inserted {
-            evicted: evicted_ids,
-        } = seen.insert(op_id)
-        {
-            evicted.extend(evicted_ids);
-        }
-    }
-    persist_seen_op_evictions(store, &evicted)?;
-
-    debug!(
-        count = seen.len(),
-        next_sequence, "hydrated seen OpId metadata from sled"
-    );
-    Ok((seen, next_sequence))
-}
-
-fn hydrate_op_log(store: &NxStore, limit: usize) -> anyhow::Result<(Vec<Op>, u64)> {
-    let limit = normalize_op_log_limit(limit);
-    let entries = store.scan_prefix(OP_LOG_STORE_PREFIX.as_bytes())?;
-    let mut ordered = Vec::with_capacity(entries.len());
-    let mut next_sequence = 0;
-
-    for (store_key, value_bytes) in entries {
-        let op_id = match logical_op_log_id(&store_key) {
-            Ok(op_id) => op_id,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid durable op log key");
-                continue;
-            }
-        };
-        let (sequence, op) = match parse_durable_op_log_value(&value_bytes) {
-            Ok(value) => value,
-            Err(e) => {
-                warn!(op_id = %op_id, error = %e, "skipping invalid durable op log entry");
-                continue;
-            }
-        };
-        if op.id.as_str() != op_id {
-            warn!(
-                key_op_id = %op_id,
-                value_op_id = %op.id,
-                "skipping durable op log entry with mismatched OpId"
-            );
-            continue;
-        }
-
-        next_sequence = next_sequence.max(sequence.saturating_add(1));
-        ordered.push((sequence, op));
-    }
-
-    ordered.sort_by_key(|(sequence, _)| *sequence);
-    let mut op_log = ordered.into_iter().map(|(_, op)| op).collect::<Vec<_>>();
-    let evicted = prune_op_log_and_return_evicted(&mut op_log, limit);
-    persist_op_log_evictions(store, &evicted)?;
-
-    debug!(
-        count = op_log.len(),
-        next_sequence, "hydrated durable op log from sled"
-    );
-    Ok((op_log, next_sequence))
-}
-
-fn hydrate_durable_gcounter_state(
-    store: &NxStore,
-    counters: &mut HashMap<String, GCounter>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(GCOUNTER_STATE_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_gcounter_state_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid durable GCounter state key");
-                continue;
-            }
-        };
-        let counter = match parse_durable_gcounter_state(&value_bytes) {
-            Ok(counter) => counter,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid durable GCounter state");
-                continue;
-            }
-        };
-
-        materialize_gcounter_value(store, &key, counter.value())?;
-        counters.insert(key, counter);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-fn hydrate_durable_pncounter_state(
-    store: &NxStore,
-    counters: &mut HashMap<String, PNCounter>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(PNCOUNTER_STATE_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_pncounter_state_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid durable PNCounter state key");
-                continue;
-            }
-        };
-        let counter = match parse_durable_pncounter_state(&value_bytes) {
-            Ok(counter) => counter,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid durable PNCounter state");
-                continue;
-            }
-        };
-
-        materialize_pncounter_value(store, &key, counter.value())?;
-        counters.insert(key, counter);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-fn hydrate_durable_lww_register_state(
-    store: &NxStore,
-    registers: &mut HashMap<String, LwwRegister>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(LWW_REGISTER_STATE_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_lww_register_state_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid durable LWW-Register state key");
-                continue;
-            }
-        };
-        let register = match parse_durable_lww_register_state(&value_bytes) {
-            Ok(register) => register,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid durable LWW-Register state");
-                continue;
-            }
-        };
-
-        materialize_lww_register_value(store, &key, register.value())?;
-        registers.insert(key, register);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-fn hydrate_durable_lww_map_state(
-    store: &NxStore,
-    maps: &mut HashMap<String, LwwMap>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(LWW_MAP_STATE_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_lww_map_state_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid durable LWW-Map state key");
-                continue;
-            }
-        };
-        let map = match parse_durable_lww_map_state(&value_bytes) {
-            Ok(map) => map,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid durable LWW-Map state");
-                continue;
-            }
-        };
-
-        materialize_lww_map_entries(store, &key, &map.entries())?;
-        maps.insert(key, map);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-fn hydrate_durable_orset_state(
-    store: &NxStore,
-    sets: &mut HashMap<String, ORSet>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(ORSET_STATE_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_orset_state_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid durable ORSet state key");
-                continue;
-            }
-        };
-        let set = match parse_durable_orset_state(&value_bytes) {
-            Ok(set) => set,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid durable ORSet state");
-                continue;
-            }
-        };
-
-        materialize_orset_elements(store, &key, &set.elements())?;
-        sets.insert(key, set);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-fn hydrate_durable_rga_state(
-    store: &NxStore,
-    rgas: &mut HashMap<String, Rga>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(RGA_STATE_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_rga_state_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid durable RGA state key");
-                continue;
-            }
-        };
-        let rga = match parse_durable_rga_state(&value_bytes) {
-            Ok(rga) => rga,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid durable RGA state");
-                continue;
-            }
-        };
-
-        materialize_rga_values(store, &key, &rga.values())?;
-        rgas.insert(key, rga);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-fn hydrate_materialized_gcounter_values(
-    store: &NxStore,
-    node_id: &NodeId,
-    counters: &mut HashMap<String, GCounter>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(GCOUNTER_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_gcounter_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid materialized GCounter key");
-                continue;
-            }
-        };
-        let value = match parse_materialized_gcounter_value(&value_bytes) {
-            Ok(value) => value,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid materialized GCounter value");
-                continue;
-            }
-        };
-        if counters.contains_key(&key) {
-            continue;
-        }
-
-        let mut counter = GCounter::new();
-        counter.increment(node_id, value);
-        counters.insert(key, counter);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-fn hydrate_materialized_pncounter_values(
-    store: &NxStore,
-    node_id: &NodeId,
-    counters: &mut HashMap<String, PNCounter>,
-) -> anyhow::Result<usize> {
-    let entries = store.scan_prefix(PNCOUNTER_STORE_PREFIX.as_bytes())?;
-    let mut hydrated = 0;
-
-    for (store_key, value_bytes) in entries {
-        let key = match logical_pncounter_key(&store_key) {
-            Ok(key) => key,
-            Err(e) => {
-                warn!(error = %e, "skipping invalid materialized PNCounter key");
-                continue;
-            }
-        };
-        let value = match parse_materialized_pncounter_value(&value_bytes) {
-            Ok(value) => value,
-            Err(e) => {
-                warn!(key = %key, error = %e, "skipping invalid materialized PNCounter value");
-                continue;
-            }
-        };
-        if counters.contains_key(&key) {
-            continue;
-        }
-
-        let mut counter = PNCounter::new();
-        if value >= 0 {
-            counter.increment(node_id, value as u64);
-        } else {
-            counter.decrement(node_id, value.unsigned_abs());
-        }
-        counters.insert(key, counter);
-        hydrated += 1;
-    }
-
-    Ok(hydrated)
-}
-
-pub(crate) fn materialize_gcounter_value(
-    store: &NxStore,
-    key: &str,
-    value: u64,
-) -> anyhow::Result<()> {
-    let store_key = materialized_gcounter_key(key);
-    store.set(&store_key, &value.to_le_bytes())?;
-    Ok(())
-}
-
-pub(crate) fn materialize_pncounter_value(
-    store: &NxStore,
-    key: &str,
-    value: i64,
-) -> anyhow::Result<()> {
-    let store_key = materialized_pncounter_key(key);
-    store.set(&store_key, &value.to_le_bytes())?;
-    Ok(())
-}
-
-pub(crate) fn materialize_lww_register_value(
-    store: &NxStore,
-    key: &str,
-    value: &[u8],
-) -> anyhow::Result<()> {
-    let store_key = materialized_lww_register_key(key);
-    store.set(&store_key, value)?;
-    Ok(())
-}
-
-pub(crate) fn materialize_lww_map_entries(
-    store: &NxStore,
-    key: &str,
-    entries: &[(String, Vec<u8>)],
-) -> anyhow::Result<()> {
-    let store_key = materialized_lww_map_key(key);
-    let entries_json = serde_json::to_vec(entries)?;
-    store.set(&store_key, &entries_json)?;
-    Ok(())
-}
-
-pub(crate) fn materialize_orset_elements(
-    store: &NxStore,
-    key: &str,
-    elements: &[String],
-) -> anyhow::Result<()> {
-    let store_key = materialized_orset_key(key);
-    let elements_json = serde_json::to_vec(elements)?;
-    store.set(&store_key, &elements_json)?;
-    Ok(())
-}
-
-pub(crate) fn materialize_rga_values(
-    store: &NxStore,
-    key: &str,
-    values: &[Vec<u8>],
-) -> anyhow::Result<()> {
-    let store_key = materialized_rga_key(key);
-    let values_json = serde_json::to_vec(values)?;
-    store.set(&store_key, &values_json)?;
-    Ok(())
-}
-
-fn parse_durable_gcounter_state(bytes: &[u8]) -> anyhow::Result<GCounter> {
-    let json = std::str::from_utf8(bytes)
-        .map_err(|e| anyhow::anyhow!("invalid UTF-8 in durable GCounter state: {e}"))?;
-    GCounter::from_json(json).map_err(|e| anyhow::anyhow!("invalid durable GCounter JSON: {e}"))
-}
-
-fn parse_durable_pncounter_state(bytes: &[u8]) -> anyhow::Result<PNCounter> {
-    let json = std::str::from_utf8(bytes)
-        .map_err(|e| anyhow::anyhow!("invalid UTF-8 in durable PNCounter state: {e}"))?;
-    PNCounter::from_json(json).map_err(|e| anyhow::anyhow!("invalid durable PNCounter JSON: {e}"))
-}
-
-fn parse_durable_lww_register_state(bytes: &[u8]) -> anyhow::Result<LwwRegister> {
-    let json = std::str::from_utf8(bytes)
-        .map_err(|e| anyhow::anyhow!("invalid UTF-8 in durable LWW-Register state: {e}"))?;
-    LwwRegister::from_json(json)
-        .map_err(|e| anyhow::anyhow!("invalid durable LWW-Register JSON: {e}"))
-}
-
-fn parse_durable_lww_map_state(bytes: &[u8]) -> anyhow::Result<LwwMap> {
-    let json = std::str::from_utf8(bytes)
-        .map_err(|e| anyhow::anyhow!("invalid UTF-8 in durable LWW-Map state: {e}"))?;
-    LwwMap::from_json(json).map_err(|e| anyhow::anyhow!("invalid durable LWW-Map JSON: {e}"))
-}
-
-fn parse_durable_orset_state(bytes: &[u8]) -> anyhow::Result<ORSet> {
-    let json = std::str::from_utf8(bytes)
-        .map_err(|e| anyhow::anyhow!("invalid UTF-8 in durable ORSet state: {e}"))?;
-    ORSet::from_json(json).map_err(|e| anyhow::anyhow!("invalid durable ORSet JSON: {e}"))
-}
-
-fn parse_durable_rga_state(bytes: &[u8]) -> anyhow::Result<Rga> {
-    let json = std::str::from_utf8(bytes)
-        .map_err(|e| anyhow::anyhow!("invalid UTF-8 in durable RGA state: {e}"))?;
-    Rga::from_json(json).map_err(|e| anyhow::anyhow!("invalid durable RGA JSON: {e}"))
-}
-
-pub(crate) fn persist_gcounter_state(
-    store: &NxStore,
-    key: &str,
-    counter: &GCounter,
-) -> anyhow::Result<()> {
-    let store_key = durable_gcounter_state_key(key);
-    let state_json = counter.to_json()?;
-    store.set(&store_key, state_json.as_bytes())?;
-    materialize_gcounter_value(store, key, counter.value())?;
-    Ok(())
-}
-
-pub(crate) fn persist_pncounter_state(
-    store: &NxStore,
-    key: &str,
-    counter: &PNCounter,
-) -> anyhow::Result<()> {
-    let store_key = durable_pncounter_state_key(key);
-    let state_json = counter.to_json()?;
-    store.set(&store_key, state_json.as_bytes())?;
-    materialize_pncounter_value(store, key, counter.value())?;
-    Ok(())
-}
-
-pub(crate) fn persist_lww_register_state(
-    store: &NxStore,
-    key: &str,
-    register: &LwwRegister,
-) -> anyhow::Result<()> {
-    let store_key = durable_lww_register_state_key(key);
-    let state_json = register.to_json()?;
-    store.set(&store_key, state_json.as_bytes())?;
-    materialize_lww_register_value(store, key, register.value())?;
-    Ok(())
-}
-
-pub(crate) fn persist_lww_map_state(
-    store: &NxStore,
-    key: &str,
-    map: &LwwMap,
-) -> anyhow::Result<()> {
-    let store_key = durable_lww_map_state_key(key);
-    let state_json = map.to_json()?;
-    store.set(&store_key, state_json.as_bytes())?;
-    materialize_lww_map_entries(store, key, &map.entries())?;
-    Ok(())
-}
-
-pub(crate) fn persist_orset_state(store: &NxStore, key: &str, set: &ORSet) -> anyhow::Result<()> {
-    let store_key = durable_orset_state_key(key);
-    let state_json = set.to_json()?;
-    store.set(&store_key, state_json.as_bytes())?;
-    materialize_orset_elements(store, key, &set.elements())?;
-    Ok(())
-}
-
-pub(crate) fn persist_rga_state(store: &NxStore, key: &str, rga: &Rga) -> anyhow::Result<()> {
-    let store_key = durable_rga_state_key(key);
-    let state_json = rga.to_json()?;
-    store.set(&store_key, state_json.as_bytes())?;
-    materialize_rga_values(store, key, &rga.values())?;
-    Ok(())
-}
-
-#[cfg(test)]
-fn persist_seen_op_batch(
-    store: &NxStore,
-    op_id: &str,
-    sequence: u64,
-    evicted: &[String],
-) -> anyhow::Result<()> {
-    let seen_key = seen_op_store_key(op_id);
-    let seen_value = sequence.to_be_bytes();
-    let delete_keys = evicted
-        .iter()
-        .map(|op_id| seen_op_store_key(op_id))
-        .collect::<Vec<_>>();
-    let delete_refs = delete_keys
-        .iter()
-        .map(|key| key.as_slice())
-        .collect::<Vec<_>>();
-
-    store.apply_batch(
-        &[(seen_key.as_slice(), seen_value.as_slice())],
-        &delete_refs,
-    )?;
-    Ok(())
-}
-
-fn collect_seen_delete_keys(evicted: &[String]) -> Vec<Vec<u8>> {
-    evicted
-        .iter()
-        .map(|op_id| seen_op_store_key(op_id))
-        .collect()
-}
-
-fn collect_op_log_delete_keys(evicted: &[Op]) -> Vec<Vec<u8>> {
-    evicted
-        .iter()
-        .map(|op| op_log_store_key(op.id.as_str()))
-        .collect()
-}
-
-fn persist_seen_op_evictions(store: &NxStore, evicted: &[String]) -> anyhow::Result<()> {
-    if evicted.is_empty() {
-        return Ok(());
-    }
-
-    let delete_keys = evicted
-        .iter()
-        .map(|op_id| seen_op_store_key(op_id))
-        .collect::<Vec<_>>();
-    let delete_refs = delete_keys
-        .iter()
-        .map(|key| key.as_slice())
-        .collect::<Vec<_>>();
-    store.apply_batch(&[], &delete_refs)?;
-    Ok(())
-}
-
-fn persist_op_log_evictions(store: &NxStore, evicted: &[Op]) -> anyhow::Result<()> {
-    if evicted.is_empty() {
-        return Ok(());
-    }
-
-    let delete_keys = collect_op_log_delete_keys(evicted);
-    let delete_refs = delete_keys
-        .iter()
-        .map(|key| key.as_slice())
-        .collect::<Vec<_>>();
-    store.apply_batch(&[], &delete_refs)?;
-    Ok(())
-}
-
-fn persist_local_ops_batch(store: &NxStore, plans: &[OpPersistencePlan]) -> anyhow::Result<()> {
-    if plans.is_empty() {
-        return Ok(());
-    }
-
-    let mut set_keys = Vec::new();
-    let mut set_values = Vec::new();
-    let mut delete_keys = Vec::new();
-
-    for plan in plans {
-        set_keys.push(seen_op_store_key(plan.op.id.as_str()));
-        set_values.push(plan.seen_sequence.to_be_bytes().to_vec());
-        set_keys.push(op_log_store_key(plan.op.id.as_str()));
-        set_values.push(encode_durable_op_log_value(plan.op_log_sequence, &plan.op)?);
-        delete_keys.extend(collect_seen_delete_keys(&plan.seen_evicted));
-        delete_keys.extend(collect_op_log_delete_keys(&plan.op_log_evicted));
-    }
-
-    let sets = set_keys
-        .iter()
-        .zip(set_values.iter())
-        .map(|(key, value)| (key.as_slice(), value.as_slice()))
-        .collect::<Vec<_>>();
-    let deletes = delete_keys
-        .iter()
-        .map(|key| key.as_slice())
-        .collect::<Vec<_>>();
-
-    store.apply_batch(&sets, &deletes)?;
-    Ok(())
-}
-
-fn persist_remote_ops_batch(
-    store: &NxStore,
-    plans: &[OpPersistencePlan],
-    updates: RemoteCrdtUpdateBatch<'_>,
-) -> anyhow::Result<()> {
-    if plans.is_empty() {
-        return Ok(());
-    }
-
-    let mut set_keys = Vec::new();
-    let mut set_values = Vec::new();
-    let mut delete_keys = Vec::new();
-    let mut changed_counter_keys = updates.counters.keys().collect::<Vec<_>>();
-    changed_counter_keys.sort();
-
-    for key in changed_counter_keys {
-        let Some(counter) = updates.counters.get(key.as_str()) else {
-            continue;
-        };
-        set_keys.push(durable_gcounter_state_key(key));
-        set_values.push(counter.to_json()?.into_bytes());
-        set_keys.push(materialized_gcounter_key(key));
-        set_values.push(counter.value().to_le_bytes().to_vec());
-    }
-
-    let mut changed_pncounter_keys = updates.pncounters.keys().collect::<Vec<_>>();
-    changed_pncounter_keys.sort();
-
-    for key in changed_pncounter_keys {
-        let Some(counter) = updates.pncounters.get(key.as_str()) else {
-            continue;
-        };
-        set_keys.push(durable_pncounter_state_key(key));
-        set_values.push(counter.to_json()?.into_bytes());
-        set_keys.push(materialized_pncounter_key(key));
-        set_values.push(counter.value().to_le_bytes().to_vec());
-    }
-
-    let mut changed_lww_register_keys = updates.lww_registers.keys().collect::<Vec<_>>();
-    changed_lww_register_keys.sort();
-
-    for key in changed_lww_register_keys {
-        let Some(register) = updates.lww_registers.get(key.as_str()) else {
-            continue;
-        };
-        set_keys.push(durable_lww_register_state_key(key));
-        set_values.push(register.to_json()?.into_bytes());
-        set_keys.push(materialized_lww_register_key(key));
-        set_values.push(register.value_bytes());
-    }
-
-    let mut changed_lww_map_keys = updates.lww_maps.keys().collect::<Vec<_>>();
-    changed_lww_map_keys.sort();
-
-    for key in changed_lww_map_keys {
-        let Some(map) = updates.lww_maps.get(key.as_str()) else {
-            continue;
-        };
-        set_keys.push(durable_lww_map_state_key(key));
-        set_values.push(map.to_json()?.into_bytes());
-        set_keys.push(materialized_lww_map_key(key));
-        set_values.push(serde_json::to_vec(&map.entries())?);
-    }
-
-    let mut changed_orset_keys = updates.orsets.keys().collect::<Vec<_>>();
-    changed_orset_keys.sort();
-
-    for key in changed_orset_keys {
-        let Some(set) = updates.orsets.get(key.as_str()) else {
-            continue;
-        };
-        set_keys.push(durable_orset_state_key(key));
-        set_values.push(set.to_json()?.into_bytes());
-        set_keys.push(materialized_orset_key(key));
-        set_values.push(serde_json::to_vec(&set.elements())?);
-    }
-
-    let mut changed_rga_keys = updates.rgas.keys().collect::<Vec<_>>();
-    changed_rga_keys.sort();
-
-    for key in changed_rga_keys {
-        let Some(rga) = updates.rgas.get(key.as_str()) else {
-            continue;
-        };
-        set_keys.push(durable_rga_state_key(key));
-        set_values.push(rga.to_json()?.into_bytes());
-        set_keys.push(materialized_rga_key(key));
-        set_values.push(serde_json::to_vec(&rga.values())?);
-    }
-
-    for plan in plans {
-        set_keys.push(seen_op_store_key(plan.op.id.as_str()));
-        set_values.push(plan.seen_sequence.to_be_bytes().to_vec());
-        set_keys.push(op_log_store_key(plan.op.id.as_str()));
-        set_values.push(encode_durable_op_log_value(plan.op_log_sequence, &plan.op)?);
-        delete_keys.extend(collect_seen_delete_keys(&plan.seen_evicted));
-        delete_keys.extend(collect_op_log_delete_keys(&plan.op_log_evicted));
-    }
-
-    let sets = set_keys
-        .iter()
-        .zip(set_values.iter())
-        .map(|(key, value)| (key.as_slice(), value.as_slice()))
-        .collect::<Vec<_>>();
-    let deletes = delete_keys
-        .iter()
-        .map(|key| key.as_slice())
-        .collect::<Vec<_>>();
-
-    store.apply_batch(&sets, &deletes)?;
-    Ok(())
-}
-
-/// Apply operations received from a remote peer as a single commit-aware batch.
-async fn apply_remote_ops(ops: &[Op], context: &RemoteOpApplyContext<'_>) -> anyhow::Result<()> {
-    if ops.is_empty() {
-        return Ok(());
-    }
-
-    let mut seen = context.seen_ops.write().await;
-    let mut log = context.op_log.write().await;
-    let mut counters = context.counters.write().await;
-    let mut pncounters = context.pncounters.write().await;
-    let mut lww_registers = context.lww_registers.write().await;
-    let mut lww_maps = context.lww_maps.write().await;
-    let mut orsets = context.orsets.write().await;
-    let mut rgas = context.rgas.write().await;
-    let mut next_seen_sequence = context.seen_ops_next_sequence.load(Ordering::Relaxed);
-    let mut next_op_log_sequence = context.op_log_next_sequence.load(Ordering::Relaxed);
-    let mut inserted_ids = Vec::new();
-    let mut inserted_ops = Vec::new();
-    let mut plans = Vec::new();
-    let mut counter_updates = HashMap::new();
-    let mut pncounter_updates = HashMap::new();
-    let mut lww_register_updates = HashMap::new();
-    let mut lww_map_updates = HashMap::new();
-    let mut orset_updates = HashMap::new();
-    let mut rga_updates = HashMap::new();
-
-    for op in ops {
-        let op_id = op.id.as_str();
-        if seen.ids.contains(op_id) || inserted_ids.iter().any(|id| id == op_id) {
-            debug!(op_id = %op.id, "skipping duplicate op");
-            continue;
-        }
-
-        let registries = RemoteCrdtRegistries {
-            counters: &counters,
-            pncounters: &pncounters,
-            lww_registers: &lww_registers,
-            lww_maps: &lww_maps,
-            orsets: &orsets,
-            rgas: &rgas,
-        };
-        let mut updates = RemoteCrdtUpdates {
-            counters: &mut counter_updates,
-            pncounters: &mut pncounter_updates,
-            lww_registers: &mut lww_register_updates,
-            lww_maps: &mut lww_map_updates,
-            orsets: &mut orset_updates,
-            rgas: &mut rga_updates,
-        };
-        apply_remote_op_to_crdt_updates(op, &registries, &mut updates);
-
-        inserted_ids.push(op_id.to_string());
-        inserted_ops.push(op.clone());
-    }
-
-    let seen_evicted = plan_seen_evictions(&seen, &inserted_ids);
-    let op_log_evicted = plan_op_log_evictions(&log, inserted_ops.len(), context.op_log_limit);
-    let last_insert_index = inserted_ops.len().saturating_sub(1);
-
-    for (index, op) in inserted_ops.iter().enumerate() {
-        plans.push(OpPersistencePlan {
-            op: op.clone(),
-            seen_sequence: next_seen_sequence,
-            op_log_sequence: next_op_log_sequence,
-            seen_evicted: if index == last_insert_index {
-                seen_evicted.clone()
-            } else {
-                Vec::new()
-            },
-            op_log_evicted: if index == last_insert_index {
-                op_log_evicted.clone()
-            } else {
-                Vec::new()
-            },
-        });
-        next_seen_sequence = next_seen_sequence.saturating_add(1);
-        next_op_log_sequence = next_op_log_sequence.saturating_add(1);
-    }
-
-    persist_remote_ops_batch(
-        context.store,
-        &plans,
-        RemoteCrdtUpdateBatch {
-            counters: &counter_updates,
-            pncounters: &pncounter_updates,
-            lww_registers: &lww_register_updates,
-            lww_maps: &lww_map_updates,
-            orsets: &orset_updates,
-            rgas: &rga_updates,
-        },
-    )?;
-
-    let applied_count = plans.len() as u64;
-    apply_seen_insertions(&mut seen, &inserted_ids, &seen_evicted);
-    log.extend(inserted_ops);
-    prune_op_log_and_return_evicted(&mut log, context.op_log_limit);
-    for (key, counter) in counter_updates {
-        counters.insert(key, counter);
-    }
-    for (key, counter) in pncounter_updates {
-        pncounters.insert(key, counter);
-    }
-    for (key, register) in lww_register_updates {
-        lww_registers.insert(key, register);
-    }
-    for (key, map) in lww_map_updates {
-        lww_maps.insert(key, map);
-    }
-    for (key, set) in orset_updates {
-        orsets.insert(key, set);
-    }
-    for (key, rga) in rga_updates {
-        rgas.insert(key, rga);
-    }
-    context
-        .seen_ops_next_sequence
-        .store(next_seen_sequence, Ordering::Relaxed);
-    context
-        .op_log_next_sequence
-        .store(next_op_log_sequence, Ordering::Relaxed);
-
-    if applied_count > 0 {
-        context.metrics.record_ops(applied_count);
-        debug!(count = applied_count, "applied remote ops batch");
-    }
-
-    Ok(())
-}
-
-fn apply_remote_op_to_crdt_updates(
-    op: &Op,
-    registries: &RemoteCrdtRegistries<'_>,
-    updates: &mut RemoteCrdtUpdates<'_>,
-) {
-    match &op.kind {
-        OpKind::GCounterIncrement { key, increment } => {
-            apply_remote_gcounter_increment(
-                op,
-                key,
-                *increment,
-                registries.counters,
-                updates.counters,
-            );
-        }
-        OpKind::PNCounterIncrement { key, .. } | OpKind::PNCounterDecrement { key, .. } => {
-            apply_remote_pncounter_op(op, key, registries.pncounters, updates.pncounters);
-        }
-        OpKind::LwwRegisterSet {
-            key,
-            value,
-            timestamp_ms,
-        } => {
-            apply_remote_lww_register_set(
-                op,
-                key,
-                value,
-                *timestamp_ms,
-                registries.lww_registers,
-                updates.lww_registers,
-            );
-        }
-        OpKind::LwwMapSet {
-            key,
-            field,
-            value,
-            timestamp_ms,
-        } => {
-            apply_remote_lww_map_set(
-                op,
-                key,
-                field,
-                value,
-                *timestamp_ms,
-                registries.lww_maps,
-                updates.lww_maps,
-            );
-        }
-        OpKind::LwwMapRemove {
-            key,
-            field,
-            timestamp_ms,
-        } => {
-            apply_remote_lww_map_remove(
-                op,
-                key,
-                field,
-                *timestamp_ms,
-                registries.lww_maps,
-                updates.lww_maps,
-            );
-        }
-        OpKind::ORSetAdd { key, element, tag } => {
-            apply_remote_orset_add(key, element, tag, registries.orsets, updates.orsets);
-        }
-        OpKind::ORSetRemove {
-            key,
-            element,
-            observed_tags,
-        } => {
-            apply_remote_orset_remove(
-                key,
-                element,
-                observed_tags,
-                registries.orsets,
-                updates.orsets,
-            );
-        }
-        OpKind::RgaInsert {
-            key,
-            id,
-            parent,
-            value,
-        } => {
-            apply_remote_rga_insert(
-                key,
-                id,
-                parent.as_deref(),
-                value,
-                registries.rgas,
-                updates.rgas,
-            );
-        }
-        OpKind::RgaDelete { key, id } => {
-            apply_remote_rga_delete(key, id, registries.rgas, updates.rgas);
-        }
-    }
-}
-
-fn apply_remote_gcounter_increment(
-    op: &Op,
-    key: &str,
-    increment: u64,
-    counters: &HashMap<String, GCounter>,
-    counter_updates: &mut HashMap<String, GCounter>,
-) {
-    let counter = counter_updates
-        .entry(key.to_string())
-        .or_insert_with(|| counters.get(key).cloned().unwrap_or_else(GCounter::new));
-    counter.increment(&op.origin, increment);
-}
-
-fn apply_remote_pncounter_op(
-    op: &Op,
-    key: &str,
-    counters: &HashMap<String, PNCounter>,
-    counter_updates: &mut HashMap<String, PNCounter>,
-) {
-    let counter = counter_updates
-        .entry(key.to_string())
-        .or_insert_with(|| counters.get(key).cloned().unwrap_or_else(PNCounter::new));
-    let _ = counter.apply_op(op);
-}
-
-fn apply_remote_lww_register_set(
-    op: &Op,
-    key: &str,
-    value: &[u8],
-    timestamp_ms: u64,
-    registers: &HashMap<String, LwwRegister>,
-    register_updates: &mut HashMap<String, LwwRegister>,
-) {
-    if let Some(register) = register_updates.get_mut(key) {
-        register.assign(value.to_vec(), timestamp_ms, op.origin.clone());
-        return;
-    }
-
-    let Some(existing) = registers.get(key) else {
-        register_updates.insert(
-            key.to_string(),
-            LwwRegister::new(value.to_vec(), timestamp_ms, op.origin.clone()),
-        );
-        return;
-    };
-
-    let mut register = existing.clone();
-    if register.assign(value.to_vec(), timestamp_ms, op.origin.clone()) {
-        register_updates.insert(key.to_string(), register);
-    }
-}
-
-fn apply_remote_lww_map_set(
-    op: &Op,
-    key: &str,
-    field: &str,
-    value: &[u8],
-    timestamp_ms: u64,
-    maps: &HashMap<String, LwwMap>,
-    map_updates: &mut HashMap<String, LwwMap>,
-) {
-    let map = map_updates
-        .entry(key.to_string())
-        .or_insert_with(|| maps.get(key).cloned().unwrap_or_else(LwwMap::new));
-    map.set(
-        field.to_string(),
-        value.to_vec(),
-        timestamp_ms,
-        op.origin.clone(),
-    );
-}
-
-fn apply_remote_lww_map_remove(
-    op: &Op,
-    key: &str,
-    field: &str,
-    timestamp_ms: u64,
-    maps: &HashMap<String, LwwMap>,
-    map_updates: &mut HashMap<String, LwwMap>,
-) {
-    let map = map_updates
-        .entry(key.to_string())
-        .or_insert_with(|| maps.get(key).cloned().unwrap_or_else(LwwMap::new));
-    map.remove(field.to_string(), timestamp_ms, op.origin.clone());
-}
-
-fn apply_remote_orset_add(
-    key: &str,
-    element: &str,
-    tag: &str,
-    sets: &HashMap<String, ORSet>,
-    set_updates: &mut HashMap<String, ORSet>,
-) {
-    let set = set_updates
-        .entry(key.to_string())
-        .or_insert_with(|| sets.get(key).cloned().unwrap_or_else(ORSet::new));
-    set.apply_add(element, tag);
-}
-
-fn apply_remote_orset_remove(
-    key: &str,
-    element: &str,
-    observed_tags: &[String],
-    sets: &HashMap<String, ORSet>,
-    set_updates: &mut HashMap<String, ORSet>,
-) {
-    let set = set_updates
-        .entry(key.to_string())
-        .or_insert_with(|| sets.get(key).cloned().unwrap_or_else(ORSet::new));
-    set.apply_remove(element, observed_tags.iter().cloned());
-}
-
-fn apply_remote_rga_insert(
-    key: &str,
-    id: &str,
-    parent: Option<&str>,
-    value: &[u8],
-    rgas: &HashMap<String, Rga>,
-    rga_updates: &mut HashMap<String, Rga>,
-) {
-    let rga = rga_updates
-        .entry(key.to_string())
-        .or_insert_with(|| rgas.get(key).cloned().unwrap_or_else(Rga::new));
-    rga.apply_insert(id.to_string(), parent.map(str::to_string), value.to_vec());
-}
-
-fn apply_remote_rga_delete(
-    key: &str,
-    id: &str,
-    rgas: &HashMap<String, Rga>,
-    rga_updates: &mut HashMap<String, Rga>,
-) {
-    let rga = rga_updates
-        .entry(key.to_string())
-        .or_insert_with(|| rgas.get(key).cloned().unwrap_or_else(Rga::new));
-    rga.apply_delete(id.to_string());
-}
+// End of main code. Test below:
 
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::runtime::{Runtime, RuntimeConfig};
+    use crate::sync_manager::{apply::*, peer::*, replication::*, storage::*};
+    use nx_net::NodeEvent;
+    use nx_sync::OpKind;
+    use std::time::Instant as StdInstant;
     use std::time::{SystemTime, UNIX_EPOCH};
     use tokio::time::{Duration, Instant, sleep};
