+++ crates/nx-core/src/sync_manager/mod.rs
@@ -1,0 +1,13 @@
+mod apply;
+mod manager;
+mod peer;
+mod replication;
+mod storage;
+mod types;
+
+pub use manager::{SyncHandle, SyncManager};
+pub use peer::PeerHealthState;
+pub(crate) use storage::{
+    persist_gcounter_state, persist_lww_map_state, persist_lww_register_state, persist_orset_state,
+    persist_pncounter_state, persist_rga_state,
+};
